+++ Src/Roflcopter.Plugin/TodoItems/TodoItemsCountProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 using JetBrains.Application.Settings;
@@ -11,10 +12,6 @@ using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Resources.Shell;
 using JetBrains.Util;
 using JetBrains.Util.DataStructures;
-#if !RS20172
-using System;
-
-#endif
 
 namespace Roflcopter.Plugin.TodoItems
 {
@@ -86,19 +83,14 @@ namespace Roflcopter.Plugin.TodoItems
                 var localTodoItemsCounts = new LocalList<TodoItemsCount>(definitions.Select(x => new TodoItemsCount(x)));
 
                 var allTodoItems = FetchAllTodoItems();
-#if !RS20172
                 var nameByIdLookup = FetchTodoMatcherNameByIdLookup();
-#endif
 
                 foreach (var todoItemChunk in allTodoItems.SelectMany(x => x))
                 foreach (var todoItem in todoItemChunk.Value) // IDEA: Parallelize?
                 {
-#if RS20172
-                    var todoItemName = todoItem.Name;
-#else
                     // IDEA: Instead store the Guid-Id in the "definition" (extend the 'CachedSettingsReader', see also R#'s TodoPatternStorage)
                     var todoItemName = nameByIdLookup[todoItem.PatternId].SingleItem();
-#endif
+
                     foreach (var todoItemsCount in localTodoItemsCounts)
                     {
                         if (todoItemName == todoItemsCount.Definition.Name)
@@ -113,11 +105,7 @@ namespace Roflcopter.Plugin.TodoItems
                 consumer.Update(todoItemsCounts);
         }
 
-#if RS20172
-        private List<ChunkHashMap<IPsiSourceFile, List<TodoItemBase>>> FetchAllTodoItems()
-#else
         private List<ChunkHashMap<IPsiSourceFile, List<ITodoItem>>> FetchAllTodoItems()
-#endif
         {
             using (_primaryTodoManager.Lock())
             using (ReadLockCookie.Create())
@@ -126,12 +114,10 @@ namespace Roflcopter.Plugin.TodoItems
             }
         }
 
-#if !RS20172
         private ILookup<Guid, string> FetchTodoMatcherNameByIdLookup()
         {
             return _primaryTodoManager.Matchers.ToLookup(x => x.Id, x => x.Name);
         }
-#endif
 
         [CanBeNull]
         private IReadOnlyCollection<TodoItemsCountDefinition> GetTodoItemsCountDefinitions()